+++ lib/nugrant/vagrant/v2/plugin.rb
@@ -1,3 +1,4 @@
+require 'nugrant/vagrant/v2/action'
 
 module Nugrant
   module Vagrant
@@ -7,6 +8,16 @@ module Nugrant
         description <<-DESC
           Plugin to define and use user specific parameters from various location inside your Vagrantfile.
         DESC
+
+        class << self
+          def provision(hook)
+            hook.before(::Vagrant::Action::Builtin::Provision, Nugrant::Vagrant::V2::Action.auto_export)
+          end
+        end
+
+        action_hook(:nugrant_provision, :machine_action_up, &method(:provision))
+        action_hook(:nugrant_provision, :machine_action_reload, &method(:provision))
+        action_hook(:nugrant_provision, :machine_action_provision, &method(:provision))
 
         command "user" do
           require_relative "command/root"
@@ -21,4 +32,3 @@ module Nugrant
     end
   end
 end
-